requirements.txt
accelerate==0.33.0
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
es-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/es_dep_news_trf-3.7.2/es_dep_news_trf-3.7.2-py3-none-any.whl
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
Flask-MySQLdb==2.0.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
ollama==0.3.0
packaging==24.1
pandas==2.2.2
preshed==3.0.9
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
thinc==8.2.5
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers==4.43.2
triton==3.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
